requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow>=10.0.0
celery==5.3.4
redis==5.0.1
django-filter==23.5
python-decouple==3.8

